pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orchestrator-mcp"
version = "0.1.0"
description = "A local-first multi-LLM orchestration MCP server with routing, safety, caching, budgeting, and observability."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "orchestrator-mcp contributors" }]
keywords = [
  "mcp",
  "model-context-protocol",
  "llm-orchestration",
  "multi-llm",
  "ollama",
  "gemini",
  "nvidia",
  "developer-tools",
  "cost-routing",
  "local-first"
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "httpx>=0.27",
  "mcp>=1.0.0",
  "pydantic>=2.7",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dashboard = [
  "streamlit>=1.43",
]
docs = [
  "mkdocs-material>=9.6",
]
dev = [
  "pytest>=8.0",
  "ruff>=0.11",
  "mypy>=1.10",
]

[project.urls]
Homepage = "https://github.com/jamiojala/orchestrator-mcp"
Repository = "https://github.com/jamiojala/orchestrator-mcp"
Issues = "https://github.com/jamiojala/orchestrator-mcp/issues"
Documentation = "https://github.com/jamiojala/orchestrator-mcp"

[project.scripts]
orchestrator-mcp = "orchestrator_mcp.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_any_generics = true
no_implicit_optional = true
check_untyped_defs = true
